Major restructuring in frontend/backend for picture backend
Added coordinator, picBackend and mainGui classes. To start with the
last class; it's job is to regroup all te signals and slots for the
coordinator to connect to. The picBackend will load and deliver the
static images. The cameraBackend is not yet written though, so this code
lingers in the drawBackground class..