CamBackend is now streaming pictures
When pressing the (new) Opencv feed button, a timed series of signals
are fired to the cambackend which returns the images to the gui. Thus,
clicking this button ones, will in fact fire it once every 20ms. This is
the simplest way to have a streaming backend. No problems with respect
to speed/load are apparent.