Camera Backend now in separate thread and streaming
To this end, I needed to implement a new ImagePacket class used for
signalling across thread boundaries. This is all working and gives a
live feed. However, many problems remain. Mainly I cannot stop and
restart the stream. It often locks up the camera and I need to replug
the connection to the avt camera to have it work again..