Resolved the VideoCapture restarting bug
It appears this was due to not opening and closing the camera in the
same 'main' thread. When releasing the camera outside the thread loop, I
could restart the streaming. However, I am not sure that the
videocapture destructor is really called but this does not seem to
bother the camera.