The fps can now be controlled by setting different delays
As qTimer can only take integer ms values, setting the fps is somewhat
misleading I think. Thus we control the timer delay in the interface.
With opencv there is no need to stop the acquisition, we can just change
the polling. Note that this will lead to less accurate timing as the
delta T can mount up to the value determined by the acquisition rate of
the camera.