Added regexsourcesink to show a movie build out of a series of \d+.png 's for instance.
If you press load Movie and select a single png, bmp or jpg image,
gigaviewer will try to fit it in a regex build up like \d+.png$ and load
all of the files in the current directory that fit that pattern. This
exposed a bug of calling 'new sourcesink' and not 'new sourcesink()'
which leaves the attributes uninitialised. This lead to strange problems
with the QString and QStringList attributes. Also forcing a full rebuild
when changing the attributes seams necessary..