changed ImageSourceSink from pure virtual to normal virtual
This was done so that non-defined methods in the children do not have to
be implemented. E.g all the camera controls for the AVT-cameras are not
supported by the other backends. Thus, dummy implementations would need
to be implemented in each other backend whereas it now simply stays in
the interface.