Added a frequency analyzer
Pressing the former max Intensity button (now 'Analyze Image'), the
image is fft'd and the max frequency component is extracted (i.e. its
period) together with the angle under which it is seen. The maximum
period that can be analyzed is amPixels/10. This is to avoid picking up
the DC component. Also, the frequency should be horizontal (as 10 first
columns of PSD are ignored.