Fixes for long intervals between images
* When the interval between two images becomes larger than a hard-coded
value of 1 second, the system will introduce skips where the camera
trigger remains smaller than 1 second but several images are skipped
from showing and recording to disk.
* Also fixed a bug in Vimba where the timeOffset was stored as an
integer which lead to negative time stamps. Now it is stored internally
as a double precision.
* Finally for compilation on windows, a console config option is
available which allows to see the qDebug output (but opens a terminal
next to the program).